Checkpoint for safe keeping
