+++ .gitignore
@@ -3,5 +3,6 @@
 .settings/
 target/
 .classpath
+.factorypath
 .project
 *.iml
